♻️ Move commit.go to git package
